AT LAST!
I got the memory to work from all aspects, I might need to explain in
person for this. Do not use the get methods anymore, you have to open by
memory.
calories class has been moved to userprofile, I needed the onCreate in
order to save to memory and the file by itself was not cutting it.
The boolean for skipping intro is working now!