Exercise page added
Added exercise page and added functionality with diary page so that it
displays exercises under exercises section. You can access it through
the add exercise button on the diary page. I also updated the equation
on the top of the diary page goal + food - exercise = remaining. Just
some notes: goal is always giving me -500 no matter what I enter for
user profile. I’m getting that value from getCalorieGoal(). Food is
initialized to 0 because i don’t know where to get that value from.
There is also a limitation on how many exercises you can add which is
3. I can increase that if you guys want. I still need to edit the
homepage so that it reflects the changes made with adding an exercise
and there needs to be a way for these values to be reset to default
whenever a new day begins. If we’re still doing the calendar, we also
need to figure out a way to create a database for these exercises.