Added Food and Workouts to Diary
Ok so I added 2 slots for each of of the three meals. You can try
clicking their different add buttons and it should add the food to that
specific view that the button you clicked was in. It updates the meals
at the top of the diary page and the calories consumed in the homepage.

Also added the workouts I did earlier to the diary. If you make a
workout and then finish it, it should add it to the diary.