Added toast for goal
I added a toast in the User Profile and Edit Profile for when a user
tries to enter a goal of losing more than 2 pounds a week.  This will
prevent the user from entering in an unhealthy goal (its typically
recommended that you lose no more than 2 pounds a week) or an outrageous
goal that would lead to a negative calorie goal number.